requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.10.8
cx-Oracle==8.3.0
mysql-connector==2.2.9
python-telegram-bot==13.9
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
tornado==6.1
tzdata==2021.5
tzlocal==4.1
XlsxWriter==3.0.2
